+++ ChangeLog.md
@@ -2,6 +2,10 @@
 
 ## NEXT
 
+  *
+
+## v1.5.3 (2011-10-21)
+
   * bundle\_without: ey.yml option allows you to specify custom bundle install --without string (list of space separated groups, replaces the default 'test development')
   * Refactor how bundler options are loaded.